v0.2.5
16 Jul 23

I created a move structure, which allows the move in correct chess notation to appear on the debug log. I think this would be useful when we have non-standard moves like castling and promotion as they are not easily encoded. I have yet to implement any special moves yet, all I have done is the simple moves for each piece with the 2 squares forward for pawn on first move.